+++ src/main/java/com/portfolio/api/repositories/ProjectRepository.java
@@ -4,8 +4,6 @@ import com.portfolio.api.models.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
 }
