+++ src/main/java/com/portfolio/api/repositories/SkillRepository.java
@@ -4,8 +4,6 @@ import com.portfolio.api.models.Skill;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface SkillRepository extends JpaRepository<Skill, Integer> {
+public interface SkillRepository extends JpaRepository<Skill, Long> {
 }
